+++ java/core/libjoynr/src/main/java/io/joynr/messaging/routing/LibJoynrMessageRouter.java
@@ -33,6 +33,7 @@ import io.joynr.messaging.MessagingSkeletonFactory;
 import io.joynr.messaging.inprocess.InProcessAddress;
 import io.joynr.runtime.ShutdownNotifier;
 import io.joynr.runtime.SystemServicesSettings;
+import io.joynr.statusmetrics.StatusReceiver;
 import joynr.ImmutableMessage;
 import joynr.exceptions.ProviderRuntimeException;
 import joynr.system.RoutingProxy;
@@ -85,7 +86,8 @@ public class LibJoynrMessageRouter extends AbstractMessageRouter {
                                  AddressManager addressManager,
                                  MulticastReceiverRegistry multicastReceiverRegistry,
                                  DelayQueue<DelayableImmutableMessage> messageQueue,
-                                 ShutdownNotifier shutdownNotifier) {
+                                 ShutdownNotifier shutdownNotifier,
+                                 StatusReceiver statusReceiver) {
         // CHECKSTYLE:ON
         super(routingTable,
               scheduler,
@@ -98,7 +100,8 @@ public class LibJoynrMessageRouter extends AbstractMessageRouter {
               addressManager,
               multicastReceiverRegistry,
               messageQueue,
-              shutdownNotifier);
+              shutdownNotifier,
+              statusReceiver);
         this.incomingAddress = incomingAddress;
     }