+++ java/jeeintegration/src/integrationTest/java/itest/io/joynr/jeeintegration/JeeMessagingEndpointTest.java
@@ -42,6 +42,7 @@ import org.slf4j.LoggerFactory;
 
 import io.joynr.jeeintegration.DefaultJoynrRuntimeFactory;
 import io.joynr.jeeintegration.JoynrIntegrationBean;
+import io.joynr.jeeintegration.JoynrStatusMetricsAggregator;
 import io.joynr.jeeintegration.ServiceProviderDiscovery;
 import io.joynr.jeeintegration.messaging.JeeMessagingApplication;
 import io.joynr.jeeintegration.messaging.JeeMessagingEndpoint;
@@ -64,7 +65,8 @@ public class JeeMessagingEndpointTest {
 						JeeIntegrationJoynrTestConfigurationProvider.class,
 						JoynrIntegrationBean.class,
 						JeeMessagingApplication.class,
-						JeeMessagingEndpoint.class
+						JeeMessagingEndpoint.class,
+						JoynrStatusMetricsAggregator.class
 						)
 				.addAsManifestResource(new File("src/main/resources/META-INF/beans.xml"));
 		// @formatter:on