+++ java/jeeintegration/src/main/java/io/joynr/jeeintegration/JoynrStatusMetrics.java
@@ -16,23 +16,27 @@
  * limitations under the License.
  * #L%
  */
-package io.joynr.messaging.mqtt.statusmetrics;
+package io.joynr.jeeintegration;
 
-public interface MqttStatusMetrics {
+/**
+ * Can be injected to receive status metrics from joynr. These metrics are useful to monitor the state of a service.
+ */
+public interface JoynrStatusMetrics {
     /**
-     * @return Returns the number of messages which were dropped since the start of the
-     * instance because an overload situation was detected.
+     * @return Returns the number of request messages which were discarded because the message queue reached its upper limit.
      */
-    long getDroppedIncomingRequestsCount();
+    int getNumDiscardedMqttRequests();
 
     /**
-     *  @return Returns whether there is a MQTT connection or not.
+     * @return Returns whether or not there is a valid connection to the MQTT broker.
      */
-    boolean isConnected();
+    boolean isConnectedToMqttBroker();
 
     /**
      * @return Returns a timestamp at which the MQTT connection became unavailable. If the connection has never been
      * established because the instance just started, the method returns the timestamp of the first connection attempt.
+     * When there is an active connection, the return value is undefined.
+     * The timestamp is provided as a Unix time (milliseconds since January 1, 1970 UTC (midnight)).
      */
-    long getOfflineSinceTimestamp();
+    long getDisconnectedFromMqttBrokerSinceTimestamp();
 }