+++ java/jeeintegration/src/test/java/test/io/joynr/jeeintegration/DefaultJoynrRuntimeFactoryTest.java
@@ -52,7 +52,9 @@ import io.joynr.messaging.MessagingQos;
 import io.joynr.messaging.NoOpRawMessagingPreprocessor;
 import io.joynr.messaging.RawMessagingPreprocessor;
 import io.joynr.messaging.mqtt.MqttClientIdProvider;
+import io.joynr.messaging.mqtt.statusmetrics.MqttStatusReceiver;
 import io.joynr.runtime.JoynrRuntime;
+import io.joynr.statusmetrics.StatusReceiver;
 import joynr.ImmutableMessage;
 import joynr.MutableMessage;
 import joynr.Request;
@@ -144,7 +146,9 @@ public class DefaultJoynrRuntimeFactoryTest {
 
         fixture = new DefaultJoynrRuntimeFactory(joynrProperties, joynrLocalDomain,
                                                  rawMessageProcessor, mqttClientIdProviderInstance,
-                                                 beanManager);
+                                                 beanManager,
+                                                 mock(StatusReceiver.class),
+                                                 mock(MqttStatusReceiver.class));
         scheduledExecutorService = mock(ScheduledExecutorService.class);
         Field executorField = DefaultJoynrRuntimeFactory.class.getDeclaredField("scheduledExecutorService");
         executorField.setAccessible(true);